Extract collaborator for ExoPlayer event listener.
